Refactor CommandManager for CreateEnvironment
