+++ src/Data/Common/index.ts
@@ -8,4 +8,5 @@ export * from "./UrlState";
 export * from "./useExpansion";
 export * from "./XmlFormatter";
 export * from "./PrimaryStateHelper";
-export * from "./PrimaryCommandManager";
+export * from "./CommandManagerWithoutEnv";
+export * from "./CommandManagerWithEnv";
