+++ src/Data/Managers/CreateEnvironment/CommandManager.ts
@@ -1,22 +1,12 @@
-import {
-  ApiHelper,
-  Command,
-  CommandManager,
-  CreateEnvironmentParams,
-  Maybe,
-} from "@/Core";
+import { ApiHelper, Command, CreateEnvironmentParams, Maybe } from "@/Core";
+import { CommandManagerWithoutEnv } from "@/Data/Common";
 
-export class CreateEnvironmentCommandManager implements CommandManager {
-  constructor(private readonly apiHelper: ApiHelper) {}
-
-  matches(command: Command.SubCommand<"CreateEnvironment">): boolean {
-    return command.kind === "CreateEnvironment";
-  }
-
-  getTrigger(
-    command: Command.SubCommand<"CreateEnvironment">
-  ): Command.Trigger<"CreateEnvironment"> {
-    return (body: CreateEnvironmentParams) => this.submit(command, body);
+export class CreateEnvironmentCommandManager extends CommandManagerWithoutEnv<"CreateEnvironment"> {
+  constructor(private readonly apiHelper: ApiHelper) {
+    super(
+      "CreateEnvironment",
+      (command) => (body) => this.submit(command, body)
+    );
   }
 
   private async submit(
@@ -25,6 +15,7 @@ export class CreateEnvironmentCommandManager implements CommandManager {
   ): Promise<Maybe.Type<Command.Error<"CreateEnvironment">>> {
     return this.apiHelper.putWithoutResponseAndEnvironment(this.getUrl(), body);
   }
+
   private getUrl(): string {
     return `/api/v2/environment`;
   }
